+++ OfficeIMO.Markdown/Converters/WordToMarkdownConverter.cs
@@ -13,7 +13,7 @@ namespace OfficeIMO.Markdown {
     /// <summary>
     /// Converts Word documents into Markdown text without relying on HTML or external tools.
     /// </summary>
-    public class WordToMarkdownConverter : IWordConverter {
+    public partial class WordToMarkdownConverter : IWordConverter {
         /// <summary>
         /// Converts a DOCX document from the provided stream into Markdown text.
         /// </summary>
@@ -58,22 +58,6 @@ namespace OfficeIMO.Markdown {
             }
 
             return sb.ToString().TrimEnd();
-        }
-
-        private static string GetParagraphText(Paragraph paragraph) {
-            StringBuilder sb = new StringBuilder();
-            foreach (var run in paragraph.Elements<Run>()) {
-                var text = run.GetFirstChild<Text>()?.Text;
-                if (string.IsNullOrEmpty(text)) {
-                    continue;
-                }
-                bool bold = run.RunProperties?.Bold != null;
-                bool italic = run.RunProperties?.Italic != null;
-                if (bold) sb.Append("**").Append(text).Append("**");
-                else if (italic) sb.Append('*').Append(text).Append('*');
-                else sb.Append(text);
-            }
-            return sb.ToString();
         }
 
         public void Convert(Stream input, Stream output, IConversionOptions options) {
